chore: Separate config vars for prune penalties (#16694)
We were using the same config variable for the slash penalty for a
pruned epoch where the data was and was not made available. This commit
adds a new env var so we can set different values for each.

Please read [contributing guidelines](CONTRIBUTING.md) and remove this
line.

For audit-related pull requests, please use the [audit PR
template](?expand=1&template=audit.md).