chore: deploy staging-ignition and testnet from CI (#16843)
Deploy staging-public, staging-ignition, and testnet as part of
deploy-staging-networks.yml.

Also the network-names within the aztec node to reflect these names of
networks, and sets their configuration.

Fixes the loadbalancer configuration for the rpc node, but comments it
out since we don’t need it currently.

Allows specifying a custom bootstrap node enr in spite of using a
network name.

Note that the chain config for `testnet` and `staging-ignition` in
`chain_l2_config.ts` are wrong. See
-
https://linear.app/aztec-labs/issue/TMNT-298/update-chain-l2-config-with-testnet-addresses
-
https://linear.app/aztec-labs/issue/TMNT-297/need-to-update-chain-l2-config-with-staging-ignition-addresses