refactor: fixing log related naming (#16882)
Fixes #15012

The naming became strange with the introduction of the concept of a
message that can be emitted either as a private log or as an offchain
message - via an offchain effect (also used to be called out-of-band).
Before we just used to have a private log.

### Changes
- Introduces "alias" for `PRIVATE_LOG_CIPHERTEXT_LEN` called
`MESSAGE_CIPHERTEXT_LEN` in `Aztec.nr`
- Renames `PRIVATE_LOG_PLAINTEXT_SIZE_IN_{BYTES,FIELDS}` →
`MESSAGE_PLAINTEXT_SIZE_IN_{BYTES,FIELDS}`