chore(docs): Replace the `dev` docs version with latest `nightly` release docs build (#16867)
This PR adds nightly docs releases.

This adds a workflow that will run at 4am UTC every day. The workflow
checks that the nightly release tag exists and builds the Aztec and
barretenberg docs at that release commit, then commits the changes back
to the `next` branch.

Note: this will mean that changes in `docs/docs` will no longer be
preview-able in the netlify docs builds.