feat!: checkpoint circuits (#16187)
Adding new checkpoint circuits as describe in the [design
doc](https://github.com/AztecProtocol/engineering-designs/pull/73).

Things different to the doc:
- The parity root is verified in the first block root in a checkpoint.
This is to speed up the overall proving time.
- The `CheckpointHeader` is the same as the previous
`ProposedBlockHeader` - we will still be validating the same values on
L1 when submitting a checkpoint.

To make the e2e tests pass without changing too many things, each
checkpoint currently only contains one block. The existing "L2Block"
class represents a block and its checkpoint. And a temporary
`L2BlockHeader` is created for it, which has all the fields required to
construct a `BlockHeader` and a `CheckpointHeader`.

The orchestrator should already work correctly for multiple blocks per
checkpoint.