fix: detect duplicate nullifier creation in oracles (#17317)
Merge after #17315.

With #17315 we can now detect and fix these things much more easily. It
turns out PXE was not checking that you had not emitted the same
nullifier twice in the same TX, and in fact the second read was getting
silently lost because duplicate writes into a `Set` are a no-op.

Had `NoteExecutionCache` had tests from the get-go this would've been
detected, but here we are. It'd not hurt to add tests now, or even to
rework it so that the `finish` function isn't as scary looking (it
mutates the object unnecessarily). But at least we can begin to use the
Noir tests to stablish some groundwork about these basic things being
caught.

Note that neither PXE nor TXE check whether nullifiers exist _in the
trees_ at creation. In PXE's case, the tx would just become invalid and
fail submission when sent.