feat: world state advances l1-l2 message tree per checkpoint (#18779)
L1->L2 messages should only be padded and inserted to the tree for the
first block in a checkpoint.
This pr also removes all references to the to-be-deprecated `L2Block` in
the world state package, except for the `blocks-added` event the
synchronizer listens to, which is still emitting the `L2Block` struct
(as part of the PublishedL2Block).