fix(avm): log gadget fuzzer merge fix (#19159)
A quick fix to the log gadget fuzzer - we now must set `.inputs` in
`ExecutionEvent`s and this was merged in just before the log fuzzer, so
I missed it (sorry!).

For some reason, after rebasing I continued to get segfaults which went
away when I just used a lighter `PureMerkleDB` in the context helper
rather than `SideEffectTrackingDB` (this is fine for the log fuzzer as
it uses the actual side effect tracker linked to the context to grab
info, but I'll probably need to revisit this for further gadget
fuzzers).

Coverage still at 100%, but does seem pretty slow to get there (possibly
upstream changes slowed this down a bit, also causing the db segfaults?)